* Fixed a bug to handle negtive object ID

+++ vsfs/common/object_storage_manager.cpp
@@ -133,7 +133,8 @@ Status ObjectStorageManager::statfs(struct statvfs* stbuf) {
 }
 
 string ObjectStorageManager::translate_path(ObjectId obj_id) const {
-  uint64_t subdir_num = obj_id % num_subdirs_;
+  uint64_t subdir_num = static_cast<uint64_t>(obj_id) %
+                        static_cast<uint64_t>(num_subdirs_);
   auto path = fs::path(base_path_) / to_string(subdir_num) / to_string(obj_id);
   return path.string();
 }
